+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## vNEXT
 - Create upgradeable voucher contract. (#4)
+- Add voucher type structure, duration, description and asset eligible. (#3)
 - Add upgradeable VoucherHub contract. (#2)
 - Init project. (#1)